Refactor Dex component import and update CSS file reference
